Refactor ToDo Page Styling (#106)
* refactor: Rearranged_css_styling

Rearranged the CSS styling for the to-do page to combine styling ffor the to-do list and to-do list item into the to-do list. Also moved the header styling into the styles component under todoGeneral Styling.

* docs: Fix_comments

Fixed comments for the media queries, to provide more information.

* refactor: Rearranged_css_styling

Rearranged CSS styling for the TodoCreateButton into todoGeneral.css as it was only 2 pieces of styling and is on the the general page.