Feature/add custom profile pictures (#109)
* feat: Add_custom_image

Adding a custom imagefrom the users local files has been implemented, however the default image isnt being displayed. Additionally, the image is scaled poorly and should be editable. Currently not being stored in the database

* feat: Profile picture upload complete

Uploading the profile picture is complete. The profile picture is not being associated with the users account

* Fix: Sonar cloud